fix(api): split replicate generation webhook handling and repair CI guards
Extract the asset/media generation branches of ReplicateWebhookController
into a dedicated ReplicateGenerationWebhookHandler service (mirroring the
existing Stripe webhook handler pattern), bringing the controller under
the 500-line runtime-complexity file limit while keeping the training
cluster inline.

Fix webhooks.service.spec.ts's mockAsset fixture, which still used the
Mongo-era parent/parentModel alias shape instead of the production
scalar-FK fields (parentBrandId, parentType).

Prune 21 stale entries from relation-alias-reads.baseline.ts left behind
by an earlier merge — those files were already fixed to read scalar FKs
(matching origin/master byte-for-byte) but the baseline still listed them
as violating, which fails check:relation-alias-reads as "stale."
