+++ apps/server/api/src/endpoints/webhooks/replicate/webhooks.replicate.controller.ts
@@ -1,23 +1,15 @@
-import { AssetsService } from '@api/collections/assets/services/assets.service';
 import { ModelRegistrationService } from '@api/collections/models/services/model-registration.service';
-import { ModelsService } from '@api/collections/models/services/models.service';
+import type { TrainingDocument } from '@api/collections/trainings/schemas/training.schema';
 import { TrainingsService } from '@api/collections/trainings/services/trainings.service';
+import { ReplicateGenerationWebhookHandler } from '@api/endpoints/webhooks/replicate/handlers/replicate-generation-webhook.handler';
 import { ReplicateWebhookService } from '@api/endpoints/webhooks/replicate/webhooks.replicate.service';
-import { WebhooksService } from '@api/endpoints/webhooks/webhooks.service';
 import { AutoSwagger } from '@api/helpers/decorators/swagger/auto-swagger.decorator';
 import { ReplicateStatus } from '@api/services/integrations/replicate/helpers/replicate.enum';
 import { NotificationsPublisherService } from '@api/services/notifications/publisher/notifications-publisher.service';
-import { resolveRelationId } from '@api/shared/utils/relation-id/relation-id.util';
-import { supportsMultipleOutputs } from '@genfeedai/constants';
-import {
-  IngredientCategory,
-  IngredientStatus,
-  ModelCategory,
-  TrainingStatus,
-} from '@genfeedai/enums';
+import { IngredientStatus, TrainingStatus } from '@genfeedai/enums';
 import { ConfigService } from '@libs/config/config.service';
 import { Public } from '@libs/decorators/public.decorator';
-import { ReplicateWebhookPayload } from '@libs/interfaces/webhook-payload.interface';
+import type { ReplicateWebhookPayload } from '@libs/interfaces/webhook-payload.interface';
 import { LoggerService } from '@libs/logger/logger.service';
 import { CallerUtil } from '@libs/utils/caller/caller.util';
 import {
@@ -41,13 +33,11 @@ export class ReplicateWebhookController {
   constructor(
     private readonly configService: ConfigService,
     private readonly loggerService: LoggerService,
-    private readonly modelsService: ModelsService,
     readonly _replicateWebhookService: ReplicateWebhookService,
-    private readonly webhooksService: WebhooksService,
     private readonly trainingsService: TrainingsService,
-    private readonly assetsService: AssetsService,
     private readonly websocketService: NotificationsPublisherService,
     private readonly modelRegistrationService: ModelRegistrationService,
+    private readonly generationWebhookHandler: ReplicateGenerationWebhookHandler,
   ) {}
 
   /**
@@ -103,327 +93,174 @@ export class ReplicateWebhookController {
       });
 
       if (existingTraining) {
-        const training = existingTraining;
-
-        if (training) {
-          if (
-            payload.status === (ReplicateStatus.COMPLETED as string) ||
-            payload.status === (ReplicateStatus.SUCCEEDED as string)
-          ) {
-            // Prefer trained model URL from output.version when present
-            const trainedModelVersion =
-              (payload?.output as Record<string, string>)?.version ||
-              payload?.version;
-            const trainingConfig =
-              training.config &&
-              typeof training.config === 'object' &&
-              !Array.isArray(training.config)
-                ? (training.config as Record<string, unknown>)
-                : {};
-
-            // Update training with completed status and model URL
-            const updatedTraining = await this.trainingsService.patch(
-              training.id,
-              {
-                config: {
-                  ...trainingConfig,
-                  model: trainedModelVersion,
-                  status: TrainingStatus.COMPLETED,
-                  trainedModelVersion,
-                },
-                stage: 'READY',
-              },
-            );
-
-            // Publish websocket event for training completion
-            await this.websocketService.publishTrainingStatus(
-              String(training.id),
-              IngredientStatus.GENERATED,
-              String(training.user),
-              {
-                externalId: payload.id,
-                model: trainedModelVersion,
-                status: TrainingStatus.COMPLETED,
-                training: updatedTraining,
-              },
-            );
-
-            // Register trained model in the model registry
-            try {
-              if (updatedTraining) {
-                await this.modelRegistrationService.createFromTraining(
-                  updatedTraining,
-                );
-              }
-            } catch (err: unknown) {
-              // Non-fatal: reconciliation job will retry
-              this.loggerService.error(
-                `Failed to register model from training ${training.id}: ${(err as Error).message}`,
-              );
-            }
-
-            this.loggerService.log(`Training completed successfully`, {
-              externalId: payload.id,
-              model: trainedModelVersion,
-            });
-          } else if (
-            payload.status === (ReplicateStatus.FAILED as string) ||
-            payload.status === (ReplicateStatus.ERROR as string)
-          ) {
-            // Update training with failed status
-            const trainingConfig =
-              training.config &&
-              typeof training.config === 'object' &&
-              !Array.isArray(training.config)
-                ? (training.config as Record<string, unknown>)
-                : {};
-            await this.trainingsService.patch(training.id, {
-              config: {
-                ...trainingConfig,
-                error: payload.error || 'Training failed',
-                status: TrainingStatus.FAILED,
-              },
-              stage: 'FAILED',
-            });
-
-            // Publish websocket event for training failure
-            await this.websocketService.publishTrainingStatus(
-              String(training.id),
-              IngredientStatus.FAILED,
-              String(training.user),
-              {
-                error: payload.error || 'Training failed',
-                externalId: payload.id,
-                status: TrainingStatus.FAILED,
-              },
-            );
-
-            // Helpful diagnostics for malformed training URL inputs
-            try {
-              const inputImages = (payload?.input as Record<string, string>)
-                ?.input_images as string | undefined;
-
-              const hasDoubleScheme =
-                typeof inputImages === 'string' &&
-                inputImages.startsWith('https://https://');
-
-              this.loggerService.error(`Training failed`, {
-                error: payload.error,
-                externalId: payload.id,
-                hint: hasDoubleScheme
-                  ? 'Detected malformed input_images URL (double https). Consider retry with corrected URL.'
-                  : undefined,
-                inputImages,
-              });
-            } catch (error: unknown) {
-              this.loggerService.warn(`Training not found for webhook`, {
-                error,
-                externalId: payload.id,
-              });
-            }
-          }
-        } else {
-          this.loggerService.warn(`Training not found for webhook`, {
-            externalId: payload.id,
-          });
-        }
+        await this.handleTrainingWebhook(existingTraining, payload);
       } else if (
         (payload.status === (ReplicateStatus.COMPLETED as string) ||
           payload.status === (ReplicateStatus.SUCCEEDED as string)) &&
         payload.model
       ) {
-        // Check if this is an asset generation first
-        const asset = await this.assetsService.findOne({
-          externalId: payload.id,
-          isDeleted: false,
-        });
-
-        if (asset) {
-          // This is an asset generation (banner/logo)
-          const output = payload.output;
-          const imageUrl =
-            typeof output === 'string'
-              ? output
-              : Array.isArray(output) && output.length > 0
-                ? output[0]
-                : null;
-
-          if (imageUrl && typeof imageUrl === 'string') {
-            await this.webhooksService.processAssetFromWebhook(
-              'replicate',
-              asset.id,
-              imageUrl,
-            );
-          } else {
-            this.loggerService.warn(
-              'Replicate webhook: no output URL for asset',
-              {
-                assetId: asset.id,
-                predictionId: payload.id,
-                status: payload.status,
-              },
-            );
-          }
-        } else {
-          // This is a regular media generation webhook (ingredient)
-          // Look up the model in the database to get its category
-          const model = await this.modelsService.findOne({
-            isDeleted: false,
-            key: payload.model,
-          });
-
-          // Map ModelCategory to IngredientCategory
-          let ingredientCategory: IngredientCategory;
-          if (model?.category) {
-            switch (model.category) {
-              case ModelCategory.VIDEO:
-                ingredientCategory = IngredientCategory.VIDEO;
-                break;
-              case ModelCategory.MUSIC:
-                ingredientCategory = IngredientCategory.MUSIC;
-                break;
-              default:
-                ingredientCategory = IngredientCategory.IMAGE;
-                break;
-            }
-          } else {
-            // Fallback: if model not found in DB, default to IMAGE
-            this.loggerService.warn(
-              `Model not found in database, defaulting to IMAGE category`,
-              { modelKey: payload.model },
-            );
-            ingredientCategory = IngredientCategory.IMAGE;
-          }
-
-          const output = payload.output;
-
-          // Check if the model supports multiple outputs
-          // Extract model key from payload (format: "owner/model-name" or ModelKey)
-          // @ts-expect-error TS2339
-          const extractedModelKey = payload.model?.split('/').pop() || '';
-          const modelSupportsMultiOutputs =
-            supportsMultipleOutputs(extractedModelKey);
-
-          if (Array.isArray(output)) {
-            const uploadTasks = output
-              .map((url, index) => {
-                if (typeof url !== 'string') {
-                  return null;
-                }
-
-                const externalId =
-                  modelSupportsMultiOutputs && output.length > 1
-                    ? `${payload.id}_${index}`
-                    : payload.id;
-
-                return this.webhooksService.processMediaFromWebhook(
-                  'replicate',
-                  ingredientCategory,
-                  externalId,
-                  url,
-                );
-              })
-              .filter((task): task is Promise<void> => Boolean(task));
-
-            if (uploadTasks.length > 0) {
-              const uploadResults = await Promise.allSettled(uploadTasks);
-              uploadResults.forEach((result, index) => {
-                if (result.status === 'rejected') {
-                  this.loggerService.error(
-                    'Replicate webhook: failed to process output',
-                    {
-                      error: result.reason,
-                      index,
-                      predictionId: payload.id,
-                    },
-                  );
-                }
-              });
-            } else {
-              this.loggerService.warn(
-                'Replicate webhook: output array contained no URLs',
-                { model: payload.model, predictionId: payload.id },
-              );
-            }
-          } else if (typeof output === 'string') {
-            await this.webhooksService.processMediaFromWebhook(
-              'replicate',
-              ingredientCategory,
-              payload.id,
-              output,
-            );
-          } else {
-            // No direct URL(s) available — log and skip
-            this.loggerService.warn(
-              'Replicate webhook: no output URLs to process',
-              {
-                hasOutput: !!output,
-                id: payload.id,
-                model: payload.model,
-                status: payload.status,
-              },
-            );
-          }
-        }
+        await this.generationWebhookHandler.handleCompleted(payload);
       } else if (
         payload.status === (ReplicateStatus.FAILED as string) ||
         payload.status === (ReplicateStatus.ERROR as string)
       ) {
-        // Check if this is a failed asset generation first
-        const asset = await this.assetsService.findOne({
-          externalId: payload.id,
-          isDeleted: false,
-        });
-
-        if (asset) {
-          this.loggerService.error(
-            'Replicate webhook: asset generation failed',
-            {
-              assetId: asset.id,
-              error: payload.error,
-              predictionId: payload.id,
-            },
-          );
-
-          // Mark asset as deleted to indicate failure
-          await this.assetsService.patch(String(asset.id), {
-            isDeleted: true,
-          });
-
-          // Notify user via websocket. Scalar FK first: `user` is a Mongo-era
-          // relation alias that this un-populated query only carries because
-          // `BaseService.normalizeDocument` back-fills it from `userId`. If it
-          // were ever absent, `String(undefined)` would yield the truthy string
-          // `"undefined"` and publish the failure event to a bogus recipient
-          // instead of skipping it.
-          const userId = resolveRelationId(asset.userId, asset.user);
-          if (userId) {
-            await this.websocketService.publishAssetStatus(
-              String(asset.id),
-              'failed',
-              userId,
-              {
-                assetId: String(asset.id),
-                category: asset.category,
-                error: payload.error || 'Asset generation failed',
-                predictionId: payload.id,
-              },
-            );
-          }
-        } else {
-          // Handle failed generation with error message for ingredients
-          await this.webhooksService.handleFailedGeneration(
-            payload.id,
-            // @ts-expect-error TS2345
-            payload.error || 'Generation failed',
-          );
-        }
+        await this.generationWebhookHandler.handleFailed(payload);
       }
     } catch (error: unknown) {
       this.loggerService.error(`${url} failed`, error);
       // Don't re-throw - webhook response already sent
       // Errors are logged but don't affect the HTTP response
       // The error is already caught and logged in handleCallback's setImmediate callback
+    }
+  }
+
+  /**
+   * Handles a webhook callback that matches an existing training record.
+   */
+  private async handleTrainingWebhook(
+    existingTraining: TrainingDocument,
+    payload: ReplicateWebhookPayload,
+  ): Promise<void> {
+    const training = existingTraining;
+
+    if (!training) {
+      this.loggerService.warn(`Training not found for webhook`, {
+        externalId: payload.id,
+      });
+      return;
+    }
+
+    if (
+      payload.status === (ReplicateStatus.COMPLETED as string) ||
+      payload.status === (ReplicateStatus.SUCCEEDED as string)
+    ) {
+      await this.handleTrainingCompleted(training, payload);
+    } else if (
+      payload.status === (ReplicateStatus.FAILED as string) ||
+      payload.status === (ReplicateStatus.ERROR as string)
+    ) {
+      await this.handleTrainingFailed(training, payload);
+    }
+  }
+
+  /**
+   * Marks a training as completed, registers the trained model, and
+   * publishes the completion event.
+   */
+  private async handleTrainingCompleted(
+    training: TrainingDocument,
+    payload: ReplicateWebhookPayload,
+  ): Promise<void> {
+    // Prefer trained model URL from output.version when present
+    const trainedModelVersion =
+      (payload?.output as Record<string, string>)?.version || payload?.version;
+    const trainingConfig =
+      training.config &&
+      typeof training.config === 'object' &&
+      !Array.isArray(training.config)
+        ? (training.config as Record<string, unknown>)
+        : {};
+
+    // Update training with completed status and model URL
+    const updatedTraining = await this.trainingsService.patch(training.id, {
+      config: {
+        ...trainingConfig,
+        model: trainedModelVersion,
+        status: TrainingStatus.COMPLETED,
+        trainedModelVersion,
+      },
+      stage: 'READY',
+    });
+
+    // Publish websocket event for training completion
+    await this.websocketService.publishTrainingStatus(
+      String(training.id),
+      IngredientStatus.GENERATED,
+      String(training.user),
+      {
+        externalId: payload.id,
+        model: trainedModelVersion,
+        status: TrainingStatus.COMPLETED,
+        training: updatedTraining,
+      },
+    );
+
+    // Register trained model in the model registry
+    try {
+      if (updatedTraining) {
+        await this.modelRegistrationService.createFromTraining(updatedTraining);
+      }
+    } catch (err: unknown) {
+      // Non-fatal: reconciliation job will retry
+      this.loggerService.error(
+        `Failed to register model from training ${training.id}: ${(err as Error).message}`,
+      );
+    }
+
+    this.loggerService.log(`Training completed successfully`, {
+      externalId: payload.id,
+      model: trainedModelVersion,
+    });
+  }
+
+  /**
+   * Marks a training as failed, publishes the failure event, and logs
+   * diagnostics for common malformed-input cases.
+   */
+  private async handleTrainingFailed(
+    training: TrainingDocument,
+    payload: ReplicateWebhookPayload,
+  ): Promise<void> {
+    // Update training with failed status
+    const trainingConfig =
+      training.config &&
+      typeof training.config === 'object' &&
+      !Array.isArray(training.config)
+        ? (training.config as Record<string, unknown>)
+        : {};
+    await this.trainingsService.patch(training.id, {
+      config: {
+        ...trainingConfig,
+        error: payload.error || 'Training failed',
+        status: TrainingStatus.FAILED,
+      },
+      stage: 'FAILED',
+    });
+
+    // Publish websocket event for training failure
+    await this.websocketService.publishTrainingStatus(
+      String(training.id),
+      IngredientStatus.FAILED,
+      String(training.user),
+      {
+        error: payload.error || 'Training failed',
+        externalId: payload.id,
+        status: TrainingStatus.FAILED,
+      },
+    );
+
+    // Helpful diagnostics for malformed training URL inputs
+    try {
+      const inputImages = (payload?.input as Record<string, string>)
+        ?.input_images as string | undefined;
+
+      const hasDoubleScheme =
+        typeof inputImages === 'string' &&
+        inputImages.startsWith('https://https://');
+
+      this.loggerService.error(`Training failed`, {
+        error: payload.error,
+        externalId: payload.id,
+        hint: hasDoubleScheme
+          ? 'Detected malformed input_images URL (double https). Consider retry with corrected URL.'
+          : undefined,
+        inputImages,
+      });
+    } catch (error: unknown) {
+      this.loggerService.warn(`Training not found for webhook`, {
+        error,
+        externalId: payload.id,
+      });
     }
   }
