+++ scripts/architecture/relation-alias-reads.baseline.ts
@@ -16,27 +16,7 @@
  */
 
 export const RELATION_ALIAS_READ_BASELINE: Readonly<Record<string, number>> = {
-  'apps/server/api/src/collections/agent-campaigns/controllers/agent-campaigns.controller.ts': 1,
-  'apps/server/api/src/collections/agent-strategies/services/agent-strategy-autopilot-execution.service.ts': 1,
-  'apps/server/api/src/collections/agent-strategies/services/agent-strategy-autopilot-performance.service.ts': 2,
-  'apps/server/api/src/collections/articles/controllers/articles.controller.ts': 2,
-  'apps/server/api/src/collections/articles/services/articles-content.service.ts': 1,
-  'apps/server/api/src/collections/bots/services/bots-livestream-delivery.service.ts': 1,
-  'apps/server/api/src/collections/content-plan-items/services/content-plan-items.service.ts': 3,
-  'apps/server/api/src/collections/images/controllers/operations/images-operations.controller.ts': 1,
-  'apps/server/api/src/collections/ingredients/controllers/ingredients-relationships.controller.ts': 1,
-  'apps/server/api/src/collections/outreach-campaigns/controllers/outreach-campaigns.controller.ts': 3,
-  'apps/server/api/src/collections/posts/controllers/operations/posts-operations.controller.ts': 5,
-  'apps/server/api/src/collections/posts/services/posts.service.ts': 1,
-  'apps/server/api/src/collections/prompts/controllers/prompts-operations.controller.ts': 7,
-  'apps/server/api/src/collections/runs/services/runs.service.ts': 2,
-  'apps/server/api/src/collections/streaks/services/streaks.service.ts': 2,
-  'apps/server/api/src/collections/trainings/controllers/trainings.controller.ts': 2,
-  'apps/server/api/src/collections/users/services/user-setup.service.ts': 1,
-  'apps/server/api/src/collections/videos/controllers/batch-interpolation.controller.ts': 1,
-  'apps/server/api/src/collections/videos/controllers/transformations/edits/videos-edits.controller.ts': 2,
-  'apps/server/api/src/collections/videos/controllers/transformations/effects/videos-effects.controller.ts': 2,
-  'apps/server/api/src/collections/videos/services/video-generation.service.ts': 3,
+  'apps/server/api/src/endpoints/webhooks/replicate/webhooks.replicate.controller.ts': 2,
   'apps/server/api/src/services/agent-campaign/campaign-winner-extraction.service.ts': 5,
   'apps/server/api/src/services/agent-campaign/content-engine.service.ts': 10,
   'apps/server/api/src/services/agent-campaign/trigger-evaluator.service.ts': 10,
